Remove dead methods, move more methods to the base class
